fix : (숫자 야구) 단위 테스트 클래스 전체 코드 이동 수정
